refactor(native): separate repository checks and delivery callbacks
Keep repository check handle publication and destruction together, separate delivery preview/confirmation endpoints, and share bounded owned UTF-8 input copying. Preserve the C ABI and existing worker ownership.
