+++ packages/agent-native-bridge/test/Spec.hs
@@ -8,6 +8,7 @@ import qualified Agent.CLI.BrowserToolsSpec as BrowserToolsSpec
 import qualified Agent.CLI.MacOS.EngineMailboxSpec as EngineMailboxSpec
 import qualified Agent.CLI.MacOS.NativeLoopEventSpec as NativeLoopEventSpec
 import qualified Agent.CLI.MacOS.RepositoryWorkersSpec as RepositoryWorkersSpec
+import qualified Agent.CLI.MacOS.RepositoryInputSpec as RepositoryInputSpec
 import qualified Agent.CLI.MacOS.TaskSchedulerSpec as TaskSchedulerSpec
 import qualified Agent.CLI.McpAdminSpec as McpAdminSpec
 import qualified Agent.CLI.ResourceAdminSpec as ResourceAdminSpec
@@ -29,6 +30,7 @@ main = hspec do
     ResourceAdminSpec.spec
     TaskSchedulerSpec.spec
     RepositoryWorkersSpec.spec
+    RepositoryInputSpec.spec
 #ifdef darwin_HOST_OS
     AccountConnectionSpec.spec
     BrowserBridgeFFISpec.spec
